bugfix (bpi): fix the update validation, finance account, account balance and tax code
- Edit -> Update now runs the same checks and "blank. Proceed?" prompt as New -> Save
- contact check no longer blocks landlines, email-only contacts, or a default that is not listed first
- ACCOUNT dropdowns list the chart of accounts instead of the payment terms
- ACCOUNT BALANCE is the total of the PENDING quotations, and PENDING now shows this partner's quotes
- Finance TAX CODE is a dropdown with a rate, like the Items tab; VAT takes the Company Setup rate
- tax code list matches spec 4.5.3 (WV-010 removed, VAT-EXEMPT added)

+++ smpc_inventory_app/Services/Setup/Model/Bpi/Bpi.cs
@@ -119,7 +119,9 @@ namespace smpc_inventory_app.Services.Setup.Model.Bpi
         public int finance_pending_branch_id { get; set; }
         public string date { get; set; }
         public string qoute_ref { get; set; }
-        public float total_price { get; set; }
+        // double, not float: this is summed into ACCOUNT BALANCE, and a float keeps only
+        // about seven digits - a 1,234,567.89 quote would total as 1,234,567.88.
+        public double total_price { get; set; }
         public string stage { get; set; }
         public string status { get; set; }
